Propagate workflow roots through wasm calls
