+++ crates/temper-server/src/odata/read.rs
@@ -927,6 +927,9 @@ async fn handle_stream_get(
         session_id: None,
         integration_config: std::collections::BTreeMap::new(),
         trace_id: String::new(),
+        workflow_root_entity_type: None,
+        workflow_root_entity_id: None,
+        workflow_run_id: None,
         http_request: None,
     };
