+++ crates/temper-wasm/src/host_trait.rs
@@ -17,6 +17,7 @@ use uuid::Uuid;
 
 use crate::metrics;
 use crate::types::WasmInvocationContext;
+use crate::workflow_headers::add_workflow_observability_headers;
 use temper_observe::wide_event::{self, EventKind, WideEvent};
 
 /// Host capabilities provided to WASM modules.
@@ -741,6 +742,8 @@ impl WasmHost for ProductionWasmHost {
             {
                 builder = builder.header("authorization", format!("Bearer {key}"));
             }
+
+            builder = add_workflow_observability_headers(builder, &filtered_headers, inv_ctx);
         }
         // determinism-ok: is_internal check uses non-deterministic URL comparison,
         // but this runs in WasmHost (not simulation), so wall-clock/network access is fine.
@@ -991,6 +994,18 @@ impl WasmHost for ProductionWasmHost {
 
         for (k, v) in &filtered_headers {
             builder = builder.header(k.as_str(), v.as_str());
+        }
+
+        let is_internal = self
+            .internal_api_base_url
+            .as_ref()
+            .is_some_and(|api_url| url.starts_with(api_url))
+            || self
+                .secrets
+                .get("temper_api_url")
+                .is_some_and(|api_url| url.starts_with(api_url.trim_end_matches('/')));
+        if is_internal && let Some(ref inv_ctx) = self.invocation_context {
+            builder = add_workflow_observability_headers(builder, &filtered_headers, inv_ctx);
         }
 
         if !headers
@@ -2024,696 +2039,5 @@ impl WasmHost for SimWasmHost {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use opentelemetry::trace::{TraceContextExt, TracerProvider as _};
-    use opentelemetry_sdk::trace::SdkTracerProvider;
-    use std::io::{Read, Write};
-    use std::net::TcpListener;
-    use std::sync::{Arc, Mutex};
-    use tracing_opentelemetry::OpenTelemetrySpanExt;
-    use tracing_subscriber::prelude::*;
-
-    #[test]
-    fn guest_metric_count_kind_is_counter() {
-        assert!(guest_metric_is_counter_kind(Some("count")));
-        assert!(guest_metric_is_counter_kind(Some("counter")));
-        assert!(!guest_metric_is_counter_kind(Some("histogram")));
-        assert!(!guest_metric_is_counter_kind(None));
-    }
-
-    /// Build a Connect frame: [flags(1)][length(4 big-endian)][payload].
-    fn make_frame(flags: u8, payload: &[u8]) -> Vec<u8> {
-        let mut frame = Vec::with_capacity(5 + payload.len());
-        frame.push(flags);
-        frame.extend_from_slice(&(payload.len() as u32).to_be_bytes());
-        frame.extend_from_slice(payload);
-        frame
-    }
-
-    #[test]
-    fn parse_single_data_frame() {
-        let payload = b"{\"stdout\":\"hello\"}";
-        let data = make_frame(0x00, payload);
-        let frames = parse_connect_frames(&data).expect("single data frame should parse");
-        assert_eq!(frames, vec!["{\"stdout\":\"hello\"}"]);
-    }
-
-    #[test]
-    fn parse_multiple_frames() {
-        let mut data = make_frame(0x00, b"{\"stdout\":\"line1\"}");
-        data.extend(make_frame(0x00, b"{\"stdout\":\"line2\"}"));
-        data.extend(make_frame(0x02, b"trailer")); // trailer frame, should be skipped
-        let frames = parse_connect_frames(&data).expect("multiple connect frames should parse");
-        assert_eq!(frames.len(), 2);
-        assert_eq!(frames[0], "{\"stdout\":\"line1\"}");
-        assert_eq!(frames[1], "{\"stdout\":\"line2\"}");
-    }
-
-    #[test]
-    fn parse_empty_input() {
-        let frames = parse_connect_frames(&[]).expect("empty input should parse");
-        assert!(frames.is_empty());
-    }
-
-    #[test]
-    fn encode_connect_json_frame_wraps_payload() {
-        let payload = "{\"hello\":\"world\"}";
-        let framed = encode_connect_json_frame(payload);
-        assert_eq!(framed[0], 0x00);
-        assert_eq!(
-            u32::from_be_bytes([framed[1], framed[2], framed[3], framed[4]]) as usize,
-            payload.len()
-        );
-        assert_eq!(&framed[5..], payload.as_bytes());
-    }
-
-    #[test]
-    fn parse_trailer_only() {
-        let data = make_frame(0x02, b"{}");
-        let frames = parse_connect_frames(&data).expect("trailer-only frame should parse");
-        assert!(frames.is_empty());
-    }
-
-    #[test]
-    fn parse_incomplete_header_errors() {
-        let result = parse_connect_frames(&[0x00, 0x00]);
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("incomplete Connect frame header")
-        );
-    }
-
-    #[test]
-    fn parse_incomplete_payload_errors() {
-        // Header says 100 bytes but only 3 available
-        let mut data = vec![0x00];
-        data.extend_from_slice(&100u32.to_be_bytes());
-        data.extend_from_slice(b"abc");
-        let result = parse_connect_frames(&data);
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("incomplete Connect frame payload")
-        );
-    }
-
-    #[test]
-    fn current_traceparent_header_prefers_active_span_context() {
-        let tracer_provider = SdkTracerProvider::builder().build();
-        let subscriber = tracing_subscriber::registry().with(
-            tracing_opentelemetry::layer().with_tracer(tracer_provider.tracer("temper-wasm-test")),
-        );
-        let _subscriber_guard = tracing::subscriber::set_default(subscriber);
-
-        let span = tracing::info_span!("wasm.reply");
-        let expected = {
-            let _guard = span.enter();
-            let span_context = tracing::Span::current()
-                .context()
-                .span()
-                .span_context()
-                .clone();
-            assert!(
-                span_context.is_valid(),
-                "test span should have an OTEL context"
-            );
-            format!(
-                "00-{}-{}-01",
-                span_context.trace_id(),
-                span_context.span_id()
-            )
-        };
-
-        let actual = span
-            .in_scope(|| current_traceparent_header(&tracing::Span::current(), None))
-            .expect("active span should produce a traceparent");
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn internal_http_call_injects_bearer_even_with_explicit_principal_headers() {
-        let listener = TcpListener::bind("127.0.0.1:0").expect("bind test listener");
-        let addr = listener.local_addr().expect("listener addr");
-        let captured = Arc::new(Mutex::new(String::new()));
-        let captured_clone = Arc::clone(&captured);
-
-        let server = std::thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept request");
-            let mut buf = [0u8; 8192];
-            let len = stream.read(&mut buf).expect("read request");
-            *captured_clone.lock().expect("capture lock") =
-                String::from_utf8_lossy(&buf[..len]).into_owned();
-
-            let body = "{}";
-            write!(
-                stream,
-                "HTTP/1.1 200 OK\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-                body.len(),
-                body
-            )
-            .expect("write response");
-        });
-
-        let mut secrets = BTreeMap::new();
-        secrets.insert("temper_api_url".to_string(), format!("http://{addr}"));
-        secrets.insert("temper_api_key".to_string(), "secret123".to_string());
-
-        let host =
-            ProductionWasmHost::new(secrets).with_invocation_context(WasmInvocationContext {
-                tenant: "default".to_string(),
-                entity_type: "Workspace".to_string(),
-                entity_id: "ws-1".to_string(),
-                trigger_action: "CreateFile".to_string(),
-                trigger_params: Value::Null,
-                entity_state: Value::Null,
-                agent_id: Some("operator".to_string()),
-                session_id: None,
-                integration_config: BTreeMap::new(),
-                trace_id: String::new(),
-                http_request: None,
-            });
-
-        let headers = vec![
-            ("X-Tenant-Id".to_string(), "default".to_string()),
-            ("x-temper-principal-kind".to_string(), "agent".to_string()),
-            ("x-temper-principal-id".to_string(), "system".to_string()),
-        ];
-
-        let (status, _) = tokio_test::block_on(host.http_call(
-            "GET",
-            &format!("http://{addr}/tdata/Directories"),
-            &headers,
-            "",
-        ))
-        .expect("internal call should succeed");
-
-        assert_eq!(status, 200);
-        server.join().expect("server thread");
-
-        let request = captured.lock().expect("capture lock").to_lowercase();
-        assert!(
-            request.contains("authorization: bearer secret123"),
-            "expected bearer token in request, got: {request}"
-        );
-        assert!(
-            request.contains("x-temper-principal-kind: agent"),
-            "expected explicit principal header to be preserved, got: {request}"
-        );
-    }
-
-    #[test]
-    fn internal_http_call_injects_bearer_from_internal_api_key_context() {
-        let listener = TcpListener::bind("127.0.0.1:0").expect("bind test listener");
-        let addr = listener.local_addr().expect("listener addr");
-        let captured = Arc::new(Mutex::new(String::new()));
-        let captured_clone = Arc::clone(&captured);
-
-        let server = std::thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept request");
-            let mut buf = [0u8; 8192];
-            let len = stream.read(&mut buf).expect("read request");
-            *captured_clone.lock().expect("capture lock") =
-                String::from_utf8_lossy(&buf[..len]).into_owned();
-
-            let body = "{}";
-            write!(
-                stream,
-                "HTTP/1.1 200 OK\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-                body.len(),
-                body
-            )
-            .expect("write response");
-        });
-
-        let mut secrets = BTreeMap::new();
-        secrets.insert("temper_api_url".to_string(), format!("http://{addr}"));
-
-        let host = ProductionWasmHost::new(secrets)
-            .with_internal_api_key(Some("ambient-secret".to_string()))
-            .with_invocation_context(WasmInvocationContext {
-                tenant: "default".to_string(),
-                entity_type: "Workspace".to_string(),
-                entity_id: "ws-1".to_string(),
-                trigger_action: "CreateFile".to_string(),
-                trigger_params: Value::Null,
-                entity_state: Value::Null,
-                agent_id: Some("operator".to_string()),
-                session_id: None,
-                integration_config: BTreeMap::new(),
-                trace_id: String::new(),
-                http_request: None,
-            });
-
-        let headers = vec![("X-Tenant-Id".to_string(), "default".to_string())];
-
-        let (status, _) = tokio_test::block_on(host.http_call(
-            "GET",
-            &format!("http://{addr}/tdata/Directories"),
-            &headers,
-            "",
-        ))
-        .expect("internal call should succeed");
-
-        assert_eq!(status, 200);
-        server.join().expect("server thread");
-
-        let request = captured.lock().expect("capture lock").to_lowercase();
-        assert!(
-            request.contains("authorization: bearer ambient-secret"),
-            "expected ambient bearer token in request, got: {request}"
-        );
-    }
-
-    #[test]
-    fn internal_http_call_uses_configured_internal_api_base_url_without_secret() {
-        let listener = TcpListener::bind("127.0.0.1:0").expect("bind test listener");
-        let addr = listener.local_addr().expect("listener addr");
-        let captured = Arc::new(Mutex::new(String::new()));
-        let captured_clone = Arc::clone(&captured);
-
-        let server = std::thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept request");
-            let mut buf = [0u8; 8192];
-            let len = stream.read(&mut buf).expect("read request");
-            *captured_clone.lock().expect("capture lock") =
-                String::from_utf8_lossy(&buf[..len]).into_owned();
-
-            let body = "{}";
-            write!(
-                stream,
-                "HTTP/1.1 200 OK\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-                body.len(),
-                body
-            )
-            .expect("write response");
-        });
-
-        let host = ProductionWasmHost::new(BTreeMap::new())
-            .with_internal_api_base_url(Some(format!("http://{addr}")))
-            .with_internal_api_key(Some("ambient-secret".to_string()))
-            .with_invocation_context(WasmInvocationContext {
-                tenant: "default".to_string(),
-                entity_type: "Session".to_string(),
-                entity_id: "ss-1".to_string(),
-                trigger_action: "WorkspaceReady".to_string(),
-                trigger_params: Value::Null,
-                entity_state: Value::Null,
-                agent_id: Some("ss-1".to_string()),
-                session_id: Some("ss-1".to_string()),
-                integration_config: BTreeMap::new(),
-                trace_id: String::new(),
-                http_request: None,
-            });
-
-        let (status, _) = tokio_test::block_on(host.http_call(
-            "GET",
-            &format!("http://{addr}/tdata/Files('file-1')/$value"),
-            &[("X-Tenant-Id".to_string(), "default".to_string())],
-            "",
-        ))
-        .expect("internal call should succeed");
-
-        assert_eq!(status, 200);
-        server.join().expect("server thread");
-
-        let request = captured.lock().expect("capture lock").to_lowercase();
-        assert!(
-            request.contains("authorization: bearer ambient-secret"),
-            "expected ambient bearer token in request, got: {request}"
-        );
-    }
-
-    #[test]
-    fn guest_log_span_attrs_include_message_and_invocation_context() {
-        let context = WasmInvocationContext {
-            tenant: "tenant-a".to_string(),
-            entity_type: "Session".to_string(),
-            entity_id: "ss-1".to_string(),
-            trigger_action: "ContextReady".to_string(),
-            trigger_params: serde_json::Value::Null,
-            entity_state: serde_json::Value::Null,
-            agent_id: Some("agent-1".to_string()),
-            session_id: Some("ss-1".to_string()),
-            integration_config: BTreeMap::new(),
-            trace_id: "0123456789abcdef0123456789abcdef".to_string(),
-            http_request: None,
-        };
-
-        let attrs = guest_log_span_attrs(
-            "error",
-            "provider failed",
-            Some(&context),
-            Some(r#"{"status":500}"#),
-        );
-
-        assert_eq!(attrs["log.severity"], "error");
-        assert_eq!(attrs["log.message"], "provider failed");
-        assert_eq!(attrs["tenant"], "tenant-a");
-        assert_eq!(attrs["entity_type"], "Session");
-        assert_eq!(attrs["entity_id"], "ss-1");
-        assert_eq!(attrs["trigger_action"], "ContextReady");
-        assert_eq!(attrs["agent_id"], "agent-1");
-        assert_eq!(attrs["gen_ai.conversation.id"], "ss-1");
-        assert_eq!(attrs["trace_id"], "0123456789abcdef0123456789abcdef");
-        assert_eq!(attrs["fields_json"], r#"{"status":500}"#);
-    }
-
-    #[test]
-    fn guest_log_span_event_is_named_for_trace_export() {
-        #[derive(Clone)]
-        struct EventCapture {
-            names: Arc<Mutex<Vec<String>>>,
-        }
-
-        impl<S> tracing_subscriber::Layer<S> for EventCapture
-        where
-            S: tracing::Subscriber,
-        {
-            fn on_event(
-                &self,
-                event: &tracing::Event<'_>,
-                _ctx: tracing_subscriber::layer::Context<'_, S>,
-            ) {
-                self.names
-                    .lock()
-                    .expect("event capture lock poisoned")
-                    .push(event.metadata().name().to_string());
-            }
-        }
-
-        let names = Arc::new(Mutex::new(Vec::new()));
-        let subscriber = tracing_subscriber::registry().with(EventCapture {
-            names: names.clone(),
-        });
-        let _subscriber_guard = tracing::subscriber::set_default(subscriber);
-        let context = WasmInvocationContext {
-            tenant: "tenant-a".to_string(),
-            entity_type: "Session".to_string(),
-            entity_id: "ss-1".to_string(),
-            trigger_action: "ContextReady".to_string(),
-            trigger_params: serde_json::Value::Null,
-            entity_state: serde_json::Value::Null,
-            agent_id: Some("agent-1".to_string()),
-            session_id: Some("ss-1".to_string()),
-            integration_config: BTreeMap::new(),
-            trace_id: "0123456789abcdef0123456789abcdef".to_string(),
-            http_request: None,
-        };
-
-        {
-            let span = tracing::info_span!("wasm.invoke");
-            let _guard = span.enter();
-            record_guest_log_span_event(
-                "error",
-                "provider failed",
-                Some(&context),
-                Some(r#"{"status":500}"#),
-            );
-        }
-
-        let names = names.lock().expect("event capture lock poisoned");
-        assert!(names.iter().any(|name| name == "wasm_guest.log"));
-    }
-
-    // --- Span-hint-header extraction (ADR-0037) ---
-
-    #[test]
-    fn split_span_hint_headers_preserves_regular_headers() {
-        let headers = vec![
-            ("content-type".to_string(), "application/json".to_string()),
-            ("authorization".to_string(), "Bearer xyz".to_string()),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert_eq!(kept.len(), 2);
-        assert_eq!(kept[0].0, "content-type");
-        assert_eq!(kept[1].0, "authorization");
-        assert!(hints.span_name.is_none());
-        assert!(hints.attributes.is_empty());
-    }
-
-    #[test]
-    fn split_span_hint_headers_extracts_span_name_case_insensitive() {
-        let headers = vec![
-            (
-                "X-Temper-Span-Name".to_string(),
-                "tool.anthropic".to_string(),
-            ),
-            ("content-type".to_string(), "application/json".to_string()),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert_eq!(hints.span_name.as_deref(), Some("tool.anthropic"));
-        assert_eq!(kept.len(), 1);
-        assert_eq!(kept[0].0, "content-type");
-    }
-
-    #[test]
-    fn split_span_hint_headers_extracts_generic_attributes() {
-        let headers = vec![
-            (
-                "X-Temper-Span-Attr-gen_ai.request.model".to_string(),
-                "claude-sonnet-4.6".to_string(),
-            ),
-            (
-                "x-temper-span-attr-tool.name".to_string(),
-                "temper_write".to_string(),
-            ),
-            ("content-type".to_string(), "application/json".to_string()),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert_eq!(kept.len(), 1);
-        assert_eq!(hints.attributes.len(), 2);
-        assert!(
-            hints
-                .attributes
-                .iter()
-                .any(|(k, v)| k == "gen_ai.request.model" && v == "claude-sonnet-4.6")
-        );
-        assert!(
-            hints
-                .attributes
-                .iter()
-                .any(|(k, v)| k == "tool.name" && v == "temper_write")
-        );
-    }
-
-    #[test]
-    fn split_span_hint_headers_strips_empty_values() {
-        let headers = vec![
-            ("X-Temper-Span-Name".to_string(), "".to_string()),
-            (
-                "X-Temper-Span-Attr-gen_ai.request.model".to_string(),
-                "".to_string(),
-            ),
-            ("X-Temper-Span-Attr-".to_string(), "ignored".to_string()),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert!(
-            kept.is_empty(),
-            "all x-temper-span-* headers should be stripped"
-        );
-        assert!(hints.span_name.is_none(), "empty name should be ignored");
-        assert!(
-            hints.attributes.is_empty(),
-            "empty key or value should be ignored"
-        );
-    }
-
-    #[test]
-    fn split_span_hint_headers_strips_reserved_unknown_prefix() {
-        // Future-proofing: unknown X-Temper-Span-* headers are stripped so they
-        // don't leak to upstream services, but we don't act on them either.
-        let headers = vec![
-            ("X-Temper-Span-Future".to_string(), "whatever".to_string()),
-            ("content-type".to_string(), "application/json".to_string()),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert_eq!(kept.len(), 1);
-        assert_eq!(kept[0].0, "content-type");
-        assert!(hints.span_name.is_none());
-        assert!(hints.attributes.is_empty());
-    }
-
-    // --- Response-capture headers (LLM Obs output) ---
-
-    #[test]
-    fn split_span_hint_headers_extracts_response_captures() {
-        let headers = vec![
-            (
-                "X-Temper-Span-Capture-Response-gen_ai.completion".to_string(),
-                "/content/0/text".to_string(),
-            ),
-            (
-                "x-temper-span-capture-response-tool.result".to_string(),
-                "/result".to_string(),
-            ),
-            ("content-type".to_string(), "application/json".to_string()),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert_eq!(kept.len(), 1);
-        assert_eq!(hints.response_captures.len(), 2);
-        assert!(
-            hints
-                .response_captures
-                .iter()
-                .any(|(k, v)| k == "gen_ai.completion" && v == "/content/0/text")
-        );
-        assert!(
-            hints
-                .response_captures
-                .iter()
-                .any(|(k, v)| k == "tool.result" && v == "/result")
-        );
-    }
-
-    #[test]
-    fn split_span_hint_headers_ignores_empty_response_capture() {
-        let headers = vec![
-            (
-                "X-Temper-Span-Capture-Response-gen_ai.completion".to_string(),
-                "".to_string(),
-            ),
-            (
-                "X-Temper-Span-Capture-Response-".to_string(),
-                "/foo".to_string(),
-            ),
-        ];
-        let (kept, hints) = split_span_hint_headers(&headers);
-        assert!(
-            kept.is_empty(),
-            "all x-temper-span-* headers should be stripped"
-        );
-        assert!(hints.response_captures.is_empty());
-    }
-
-    #[test]
-    fn truncate_for_span_attr_passes_through_short_values() {
-        let s = "hello world";
-        assert_eq!(truncate_for_span_attr(s), s);
-    }
-
-    #[test]
-    fn truncate_for_span_attr_truncates_long_values_on_utf8_boundary() {
-        // Build a string just over the budget with a multi-byte char near the cut.
-        let mut s = "a".repeat(MAX_RESPONSE_CAPTURE_BYTES - 1);
-        s.push('🎉'); // 4 bytes, starts at MAX - 1, extends past MAX.
-        s.push_str("extra");
-        let truncated = truncate_for_span_attr(&s);
-        assert!(truncated.ends_with("…[truncated]"));
-        assert!(
-            truncated.len() <= MAX_RESPONSE_CAPTURE_BYTES + "…[truncated]".len(),
-            "truncated length {} exceeded expected cap",
-            truncated.len()
-        );
-        // Must remain valid UTF-8 (call site requires it for span attrs).
-        let _ = std::str::from_utf8(truncated.as_bytes()).expect("truncated must be valid utf-8");
-    }
-
-    #[test]
-    fn apply_response_captures_is_safe_when_body_is_not_json() {
-        // Should not panic on malformed JSON; call site passes through raw body.
-        let span = tracing::Span::none();
-        apply_response_captures(
-            &span,
-            "this is not JSON",
-            &[(
-                "gen_ai.completion".to_string(),
-                "/content/0/text".to_string(),
-            )],
-        );
-    }
-
-    #[test]
-    fn apply_response_captures_is_safe_when_pointer_misses() {
-        let span = tracing::Span::none();
-        apply_response_captures(
-            &span,
-            r#"{"nope": "not here"}"#,
-            &[(
-                "gen_ai.completion".to_string(),
-                "/content/0/text".to_string(),
-            )],
-        );
-    }
-
-    #[test]
-    fn default_http_call_batch_runs_requests_concurrently() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-        use std::time::Duration;
-
-        struct ConcurrentBatchHost {
-            current_in_flight: AtomicUsize,
-            max_in_flight: AtomicUsize,
-        }
-
-        #[async_trait]
-        impl WasmHost for ConcurrentBatchHost {
-            async fn http_call(
-                &self,
-                method: &str,
-                url: &str,
-                _headers: &[(String, String)],
-                _body: &str,
-            ) -> Result<(u16, String), String> {
-                let in_flight = self.current_in_flight.fetch_add(1, Ordering::SeqCst) + 1;
-                self.max_in_flight.fetch_max(in_flight, Ordering::SeqCst);
-                tokio::time::sleep(Duration::from_millis(25)).await;
-                self.current_in_flight.fetch_sub(1, Ordering::SeqCst);
-                Ok((200, format!("{method} {url}")))
-            }
-
-            async fn http_call_binary(
-                &self,
-                _method: &str,
-                _url: &str,
-                _headers: &[(String, String)],
-                _body: &[u8],
-            ) -> Result<(u16, Vec<u8>), String> {
-                Err("binary calls not used in this test".to_string())
-            }
-
-            fn get_secret(&self, _key: &str) -> Result<String, String> {
-                Err("secrets not used in this test".to_string())
-            }
-
-            fn log(&self, _level: &str, _message: &str) {}
-        }
-
-        let host = ConcurrentBatchHost {
-            current_in_flight: AtomicUsize::new(0),
-            max_in_flight: AtomicUsize::new(0),
-        };
-
-        let responses = tokio_test::block_on(host.http_call_batch(&[
-            HttpBatchRequest {
-                method: "GET".to_string(),
-                url: "https://example.com/a".to_string(),
-                headers: vec![],
-                body: String::new(),
-            },
-            HttpBatchRequest {
-                method: "GET".to_string(),
-                url: "https://example.com/b".to_string(),
-                headers: vec![],
-                body: String::new(),
-            },
-            HttpBatchRequest {
-                method: "GET".to_string(),
-                url: "https://example.com/c".to_string(),
-                headers: vec![],
-                body: String::new(),
-            },
-        ]))
-        .expect("batch call should succeed");
-
-        assert_eq!(responses.len(), 3);
-        assert_eq!(responses[0].body, "GET https://example.com/a");
-        assert!(
-            host.max_in_flight.load(Ordering::SeqCst) > 1,
-            "default batch implementation should overlap independent requests"
-        );
-    }
-}
+#[cfg(test)]
+mod tests;
