+++ crates/temper-wasm/tests/e2e_invoke.rs
@@ -26,6 +26,9 @@ fn build_context() -> WasmInvocationContext {
         session_id: None,
         integration_config: std::collections::BTreeMap::new(),
         trace_id: String::new(),
+        workflow_root_entity_type: None,
+        workflow_root_entity_id: None,
+        workflow_run_id: None,
         http_request: None,
     }
 }
